Приложение отправляет в сеть, и прогон по стенду поймал две поломки
Сквозной путь против живого стенда пройден целиком после всех сегодняшних правок:
`Accepted(serverMessageId=7874446234362675763)`, конверт 1672 байта, «текст сошёлся
целиком» — причём принятый текст читается ТЕПЕРЬ ИЗ БАЗЫ, то есть проверена и вчерашняя
починка записи разобранного тела.

Сделано:

- `Отправитель` в приложении: для каждой переписки с ожидающими сообщениями берёт ЕЁ ключ
  эпохи escrow, проверяет подпись анклава ДО всякого шифрования, собирает адресатов
  (устройства собеседника И свои остальные — без своих второе устройство не прочитает
  отправленное с первого) и отдаёт насосу. Отправка идёт от ИЗМЕНЕНИЙ в базе, а не по
  таймеру; таймер остался только ради сроков повторов, и это названо вслух;
- новая переписка по номеру: случай использования в `domain-chat` (9 тестов), справочник
  `UsersApi` в `core-network`, вычисление `chat_id` переходником — Domain криптографию не
  видит. Главное правило: `chat_id` личной переписки ВЫВОДИТСЯ из пары, а не выдаётся
  сервером. Свой номер — отдельный случай и в личную переписку не превращается;
- «такого номера в TIMA нет» — не отказ, а предложение позвать человека;
- поле ввода переехало из `feature-auth` в дизайн-систему: на втором экране с вводом его
  пришлось бы повторить, а с этого начинается расхождение;
- навигация верхнего уровня — три случая и ни одного стека: из подокна выходят «назад», из
  списка открывают одно.

ДВЕ ПОЛОМКИ, НАЙДЕННЫЕ ПРОГОНОМ:

1. ПОЛОВИНА СООБЩЕНИЙ НЕ УХОДИЛА, А ПРОГОН БЫЛ ЗЕЛЁНЫМ. `message_id` выводится из первых
   8 байт UUID. На проводе поле `uint64` и проходит прекрасно; сервер хранит его в
   `bigint` — ЗНАКОВОМ int64, — и Postgres отвечает ошибкой кодирования, а сервер `500`.
   Старший бит случайного UUID установлен в половине случаев: поэтому предыдущий прогон
   был зелёным, а сегодняшний нет. Бит сбрасывается, инвариант записан в
   `crypto-invariants.mdc`, проверка гоняет тысячу ключей — случайность здесь не помощник,
   а причина, по которой поломку не увидели раньше.

   Отдельно: сервер отвечает `500` на входные данные клиента. Клиент таких чисел больше не
   посылает, но правильный ответ на негодное поле — `400`. Это отдельная серверная задача,
   не про контракт, а про устойчивость.

2. `RETRY` НЕ НЕС НИ СЛОВА О ПРИЧИНЕ. 401 с просроченным токеном, 500 на сервере и
   неизвестный ответ посредника выглядели одинаково: `Retry(afterMs=1000)`. Прогон упёрся
   ровно в это — «не ушло» видно, а искать причину нечем; пришлось идти в логи сервера по
   SSH. Причина добавлена строкой и только для диагностики: на решение очереди она не
   влияет. Найдись она раньше, находка 1 стоила бы одного прогона, а не трёх.

Заодно: `String.format` в общем коде нет и быть не может — он только на JVM. Поймал
компилятор при сборке под Apple, как и задумано с К1.

411 тестов на JVM, полная сборка и архитектурные правила зелёные.

+++ client/core/core-database/src/commonTest/kotlin/io/tima/core/database/SqlOutboxStoreTest.kt
@@ -149,5 +149,5 @@ class SqlOutboxStoreTest {
      * Настоящая подпись просит переписку: ключ эпохи escrow у каждой свой.
      */
     private fun Outbox.sealNext(эпоха: Int, seal: (OutboxEntry) -> ByteArray): OutboxEntry? =
-        this.sealNext("chat-1", эпоха, seal)
+        this.sealNext("chat-1", эпоха.toLong(), seal)
 }
